Refactor report related packaging
