+++ interfaces/interfaces.go
@@ -14,12 +14,7 @@
 
 package interfaces
 
-import (
-	"fmt"
-
-	"github.com/Mirantis/k8s-AppController/client"
-	"github.com/Mirantis/k8s-AppController/human"
-)
+import "github.com/Mirantis/k8s-AppController/client"
 
 // BaseResource is an interface for AppController supported resources
 type BaseResource interface {
@@ -38,29 +33,6 @@ type DependencyReport struct {
 	Percentage int
 	Needed     int
 	Message    string
-}
-
-// AsHuman returns a human-readable representation of the report as a slice
-func (d DependencyReport) AsHuman(indent int) []string {
-	var blocksStr, percStr string
-	if d.Blocks {
-		blocksStr = "BLOCKS"
-	} else {
-		blocksStr = "DOESN'T BLOCK"
-	}
-	if d.Percentage == 100 {
-		percStr = ""
-	} else {
-		percStr = fmt.Sprintf("%d%%/%d%%", d.Percentage, d.Needed)
-	}
-	ret := []string{
-		fmt.Sprintf("Dependency: %s", d.Dependency),
-		blocksStr,
-	}
-	if percStr != "" {
-		ret = append(ret, percStr)
-	}
-	return human.Indent(indent, ret)
 }
 
 // Resource is an interface for a base resource that implements getting dependency reports
